requirements.txt
praw>=7.7.0
requests>=2.31.0
Pillow>=10.0.0
tqdm>=4.66.0
gallery-dl>=1.27.0
